chore(core): move create vault v2
