+++ mobile-app/app/screens/AppNavigator/screens/Loans/screens/CreateVaultScreen.tsx
@@ -1,29 +1,40 @@
 import { View } from "@components";
-import { Button } from "@components/Button";
-import {
-  ThemedScrollView,
-  ThemedSectionTitle,
-  ThemedText,
-} from "@components/themed";
+import { ThemedScrollViewV2, ThemedTextV2 } from "@components/themed";
 import { StackScreenProps } from "@react-navigation/stack";
 import { useWhaleApiClient } from "@shared-contexts/WhaleContext";
 import { tailwind } from "@tailwind";
 import { translate } from "@translations";
-import { useEffect, useState } from "react";
+import { Dispatch, useEffect, useState } from "react";
 import { LoanScheme } from "@defichain/whale-api-client/dist/api/loan";
 import BigNumber from "bignumber.js";
-import { useLogger } from "@shared-contexts/NativeLoggingProvider";
+import {
+  NativeLoggingProps,
+  useLogger,
+} from "@shared-contexts/NativeLoggingProvider";
 import { useSelector } from "react-redux";
-import { ascColRatioLoanScheme, fetchLoanSchemes } from "@store/loans";
+import {
+  ascColRatioLoanScheme,
+  fetchLoanSchemes,
+  fetchVaults,
+} from "@store/loans";
 import { RootState } from "@store";
-import { hasTxQueued } from "@store/transaction_queue";
+import { hasTxQueued, transactionQueue } from "@store/transaction_queue";
 import { hasTxQueued as hasBroadcastQueued } from "@store/ocean";
-import { DFITokenSelector, DFIUtxoSelector } from "@store/wallet";
-import { ConversionInfoText } from "@components/ConversionInfoText";
-import { InfoTextLink } from "@components/InfoTextLink";
+import { DFIUtxoSelector } from "@store/wallet";
 import { queueConvertTransaction } from "@hooks/wallet/Conversion";
 import { useAppDispatch } from "@hooks/useAppDispatch";
-import { LoanSchemeOptions } from "../components/LoanSchemeOptions";
+import { LoanSchemeOptionsV2 } from "@screens/AppNavigator/screens/Loans/components/LoanSchemeOptionsV2";
+import { ButtonV2 } from "@components/ButtonV2";
+import {
+  ConversionStatus,
+  CreateVaultSummary,
+} from "@screens/AppNavigator/screens/Loans/components/CreateVaultSummary";
+import { EnvironmentNetwork } from "@environment";
+import { useNetworkContext } from "@shared-contexts/NetworkContext";
+import { onTransactionBroadcast } from "@api/transaction/transaction_commands";
+import { WhaleWalletAccount } from "@defichain/whale-api-wallet";
+import { CTransactionSegWit } from "@defichain/jellyfish-transaction";
+import { useWalletContext } from "@shared-contexts/WalletContext";
 import { LoanParamList } from "../LoansNavigator";
 
 type Props = StackScreenProps<LoanParamList, "CreateVaultScreen">;
@@ -31,6 +42,8 @@ type Props = StackScreenProps<LoanParamList, "CreateVaultScreen">;
 export function CreateVaultScreen({ navigation, route }: Props): JSX.Element {
   const dispatch = useAppDispatch();
   const client = useWhaleApiClient();
+  const { network } = useNetworkContext();
+  const { address } = useWalletContext();
   const loanSchemes = useSelector((state: RootState) =>
     ascColRatioLoanScheme(state.loans)
   );
@@ -38,10 +51,6 @@ export function CreateVaultScreen({ navigation, route }: Props): JSX.Element {
     (state: RootState) => state.loans.hasFetchedLoanSchemes
   );
   const logger = useLogger();
-  const [fee, setFee] = useState<BigNumber>(new BigNumber(0.0001));
-  const [selectedLoanScheme, setSelectedLoanScheme] = useState<
-    LoanScheme | undefined
-  >(route.params?.loanScheme);
   const hasPendingJob = useSelector((state: RootState) =>
     hasTxQueued(state.transactionQueue)
   );
@@ -51,18 +60,28 @@ export function CreateVaultScreen({ navigation, route }: Props): JSX.Element {
   const DFIUtxo = useSelector((state: RootState) =>
     DFIUtxoSelector(state.wallet)
   );
-  const DFIToken = useSelector((state: RootState) =>
-    DFITokenSelector(state.wallet)
+
+  const RESERVE_AMOUNT = 2.1;
+  const [fee, setFee] = useState<BigNumber>(new BigNumber(0.0001));
+  const [selectedLoanScheme, setSelectedLoanScheme] = useState<
+    LoanScheme | undefined
+  >(route.params?.loanScheme);
+  const [conversionStatus, setConversionStatus] = useState<ConversionStatus>(
+    new BigNumber(RESERVE_AMOUNT).gt(DFIUtxo.amount)
+      ? ConversionStatus.Required
+      : ConversionStatus.Not_Required
   );
-  const isConversionRequired = new BigNumber(2.1).gt(DFIUtxo.amount);
-  const goToVaultsFaq = (): void => {
-    navigation.navigate({
-      name: "LoansFaq",
-      params: {
-        activeSessions: [2],
-      },
-    });
-  };
+  const [conversionAmount, setConversionAmount] = useState<
+    BigNumber | undefined
+  >();
+  const [isOnPage, setIsOnPage] = useState<boolean>(true);
+
+  const vaultFee = new BigNumber(
+    network === EnvironmentNetwork.MainNet ||
+    network === EnvironmentNetwork.TestNet
+      ? 2
+      : 1
+  );
 
   const onSubmit = async (): Promise<void> => {
     if (
@@ -73,40 +92,55 @@ export function CreateVaultScreen({ navigation, route }: Props): JSX.Element {
       return;
     }
 
-    if (isConversionRequired) {
+    if (conversionStatus === ConversionStatus.Required) {
+      const convertAmount = new BigNumber(RESERVE_AMOUNT).minus(DFIUtxo.amount);
       queueConvertTransaction(
         {
           mode: "accountToUtxos",
-          amount: new BigNumber(2.1).minus(DFIUtxo.amount),
+          amount: convertAmount,
         },
         dispatch,
         () => {
-          navigation.navigate({
-            name: "ConfirmCreateVaultScreen",
-            params: {
-              loanScheme: selectedLoanScheme,
-              fee: fee,
-              conversion: {
-                DFIUtxo,
-                DFIToken,
-                isConversionRequired,
-                conversionAmount: new BigNumber(2.1).minus(DFIUtxo.amount),
-              },
-            },
-          });
+          setConversionAmount(convertAmount);
+          setConversionStatus(ConversionStatus.Processing);
+        },
+        logger,
+        () => {
+          setConversionStatus(ConversionStatus.Completed);
+        }
+      );
+    } else {
+      if (hasPendingJob || hasPendingBroadcastJob) {
+        return;
+      }
+      await createVault(
+        {
+          address: address,
+          loanScheme: selectedLoanScheme,
+        },
+        dispatch,
+        () => {
+          onTransactionBroadcast(isOnPage, navigation.dispatch);
+        },
+        () => {
+          dispatch(
+            fetchVaults({
+              address,
+              client,
+            })
+          );
         },
         logger
       );
-    } else {
-      navigation.navigate({
-        name: "ConfirmCreateVaultScreen",
-        params: {
-          loanScheme: selectedLoanScheme,
-          fee: fee,
-        },
-      });
     }
   };
+
+  useEffect(() => {
+    setIsOnPage(true);
+    return () => {
+      setIsOnPage(false);
+    };
+  }, []);
 
   useEffect(() => {
     dispatch(fetchLoanSchemes({ client }));
@@ -119,38 +153,38 @@ export function CreateVaultScreen({ navigation, route }: Props): JSX.Element {
       .catch(logger.error);
   }, []);
 
+  useEffect(() => {
+    if (
+      conversionStatus === ConversionStatus.Processing ||
+      conversionStatus === ConversionStatus.Completed
+    ) {
+      return;
+    }
+
+    const needsConvert = new BigNumber(RESERVE_AMOUNT).gt(DFIUtxo.amount);
+    setConversionStatus(
+      needsConvert ? ConversionStatus.Required : ConversionStatus.Not_Required
+    );
+    if (needsConvert) {
+      setConversionAmount(undefined);
+    }
+  }, [DFIUtxo]);
+
   return (
-    <ThemedScrollView
+    <ThemedScrollViewV2
       testID="create_vault_screen"
-      contentContainerStyle={tailwind("py-6 pb-8 px-4")}
+      contentContainerStyle={tailwind("py-8 px-5")}
     >
-      <ThemedSectionTitle
-        light={tailwind("text-gray-900")}
-        dark={tailwind("text-gray-50")}
-        text={translate(
-          "screens/CreateVaultScreen",
-          "Choose loan scheme for your vault"
-        )}
-        style={tailwind("mb-2 font-semibold text-lg")}
-      />
-      <ThemedText
-        light={tailwind("text-gray-700")}
-        dark={tailwind("text-gray-200")}
-        style={tailwind("text-sm")}
+      <ThemedTextV2
+        style={tailwind("text-base font-normal-v2 px-3 text-center")}
       >
         {translate(
           "screens/CreateVaultScreen",
-          "Loan scheme of your vault determines the required collateralization of your vault for loans."
+          "Select a loan scheme for your vault."
         )}
-      </ThemedText>
-      <View style={tailwind("mt-2 mb-6")}>
-        <InfoTextLink
-          onPress={goToVaultsFaq}
-          text="Learn more about vaults and loan schemes"
-          testId="empty_vault_learn_more"
-        />
-      </View>
-      <LoanSchemeOptions
+      </ThemedTextV2>
+
+      <LoanSchemeOptionsV2
         loanSchemes={loanSchemes}
         isLoading={!hasFetchedLoanSchemes}
         selectedLoanScheme={selectedLoanScheme}
@@ -158,43 +192,112 @@ export function CreateVaultScreen({ navigation, route }: Props): JSX.Element {
           setSelectedLoanScheme(scheme)
         }
       />
-      <ThemedText
-        light={tailwind("text-gray-500")}
-        dark={tailwind("text-gray-400")}
-        style={tailwind("text-center text-xs mb-10")}
-      >
-        {translate(
-          "screens/CreateVaultScreen",
-          "Keep note of your selected collateralization ratio for your vault to sustain the loans within it."
-        )}
-      </ThemedText>
 
-      {isConversionRequired && (
-        <View style={tailwind("mt-4 mb-6")}>
-          <ConversionInfoText />
-        </View>
-      )}
-      <Button
+      {selectedLoanScheme !== undefined &&
+        (conversionStatus !== ConversionStatus.Required ||
+          conversionAmount !== undefined) && (
+          <CreateVaultSummary
+            transactionFee={fee}
+            vaultFee={vaultFee}
+            convertAmount={conversionAmount}
+            conversionStatus={conversionStatus}
+          />
+        )}
+
+      <ButtonActionMessage
+        isConversionRequired={conversionStatus === ConversionStatus.Required}
+        hasSelectedLoanScheme={selectedLoanScheme !== undefined}
+      />
+      <ButtonV2
         disabled={
           selectedLoanScheme === undefined ||
           hasPendingJob ||
           hasPendingBroadcastJob
         }
-        label={translate("screens/CreateVaultScreen", "CONTINUE")}
+        label={translate(
+          "screens/CreateVaultScreen",
+          conversionStatus === ConversionStatus.Required
+            ? "Continue"
+            : "Create vault"
+        )}
         onPress={onSubmit}
-        margin="m-0 mb-2"
+        styleProps="mt-0 mx-7"
         testID="create_vault_submit_button"
       />
-      <ThemedText
-        light={tailwind("text-gray-500")}
-        dark={tailwind("text-gray-400")}
-        style={tailwind("text-center text-xs")}
-      >
-        {translate(
-          "screens/CreateVaultScreen",
-          "Confirm your vault details in next screen"
-        )}
-      </ThemedText>
-    </ThemedScrollView>
+    </ThemedScrollViewV2>
   );
+}
+
+function ButtonActionMessage(props: {
+  isConversionRequired: boolean;
+  hasSelectedLoanScheme: boolean;
+}): JSX.Element {
+  return (
+    <View style={tailwind("pt-12 px-12")}>
+      {props.hasSelectedLoanScheme && (
+        <ThemedTextV2
+          style={tailwind("text-xs font-normal-v2 pb-5 text-center")}
+          light={tailwind("text-mono-light-v2-500")}
+          dark={tailwind("text-mono-dark-v2-500")}
+          testID="action_message"
+        >
+          {translate(
+            "screens/CreateVaultScreen",
+            props.isConversionRequired
+              ? "By continuing, the required amount of DFI will be converted"
+              : "Monitor your vault’s collateralization to prevent liquidation."
+          )}
+        </ThemedTextV2>
+      )}
+    </View>
+  );
+}
+
+interface VaultForm {
+  loanScheme: LoanScheme;
+  address: string;
+}
+
+async function createVault(
+  { loanScheme }: VaultForm,
+  dispatch: Dispatch<any>,
+  onBroadcast: () => void,
+  onConfirmation: () => void,
+  logger: NativeLoggingProps
+): Promise<void> {
+  try {
+    const signer = async (
+      account: WhaleWalletAccount
+    ): Promise<CTransactionSegWit> => {
+      const script = await account.getScript();
+      const builder = account.withTransactionBuilder();
+      const signed = await builder.vault.createVault(
+        {
+          ownerAddress: script,
+          schemeId: loanScheme.id,
+        },
+        script
+      );
+      return new CTransactionSegWit(signed);
+    };
+
+    dispatch(
+      transactionQueue.actions.push({
+        sign: signer,
+        title: translate("screens/CreateVaultScreen", "Creating vault"),
+        drawerMessages: {
+          preparing: translate(
+            "screens/OceanInterface",
+            "Preparing to create vault…"
+          ),
+          waiting: translate("screens/OceanInterface", "Creating vault"),
+          complete: translate("screens/OceanInterface", "Vault created"),
+        },
+        onBroadcast,
+        onConfirmation,
+      })
+    );
+  } catch (e) {
+    logger.error(e);
+  }
 }
